+++ src/layout/vpx_video_producer.cpp
@@ -23,9 +23,8 @@ VPXVideoProducer::VPXVideoProducer(const hisui::Config& t_config,
                                    const std::uint64_t timescale)
     : VideoProducer({.show_progress_bar = t_config.show_progress_bar}) {
   auto resolution = t_metadata.getResolution();
-  hisui::video::VPXEncoderConfig vpx_config(
-      static_cast<std::uint32_t>(resolution.width),
-      static_cast<std::uint32_t>(resolution.height), t_config);
+  hisui::video::VPXEncoderConfig vpx_config(resolution.width, resolution.height,
+                                            t_config);
 
   // TODO(haruyama): composer の設定
